refactor(router): move interfaces and types to interfaces
